src/harp/input.py
"""
Keyboard emulation for Wayland.
"""

import uinput


class WaylandTyper:
    """
    Emulates a physical keyboard to type text in Wayland environments.
    """

    def __init__(self) -> None:
        """
        Initializes the WaylandTyper device with US keyboard capabilities.
        """
        # Placeholder for initializing a uinput.Device with standard US keys.
        self.device: uinput.Device | None = None

    def type_text(self, text: str) -> None:
        """
        Emulates keystrokes for the provided text.

        Args:
            text: The text to be typed into the active window.
        """
        # Placeholder for converting characters to uinput key events.
        pass
