pyproject.toml
[project]
name = "harp"
version = "0.1.0"
description = "A background daemon for Wayland that transcribes audio and types results."
readme = "README.md"
authors = [
    { name = "Alejandro Piad", email = "apiad@apiad.net" }
]
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.2",
    "openai>=1.65.0",
    "pydantic-settings>=2.13.1",
    "python-uinput>=1.0.1",
    "sounddevice>=0.5.5",
    "typer>=0.24.1",
]

[tool.ruff]
exclude = [".gemini"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.4",
]
